A spot of refactoring
